Add support for virtual autochanger devices
Standard storage Devices are quite restrictive in what they allow
concurrently. The built-in virtual autochanger allows multiple jobs to
run at the same time to different volumes, and for restores to be done
at the same time as backups.

This commit adds a new defined type for `autochanger_device`, which
allows management of an autochanger with as many underlying devices as
desired.